Fix production build
